+++ js/filter-images.js
@@ -1,0 +1,1 @@
+// temporarily empty module for the following functions
